refactor the ROSH risk stuff into a shared presenter and view class
