+++ server/routes/referrals/riskInformationView.ts
@@ -1,10 +1,12 @@
 import RiskInformationPresenter from './riskInformationPresenter'
 import ViewUtils from '../../utils/viewUtils'
-import { DetailsArgs, TableArgs, TagArgs, TextareaArgs } from '../../utils/govukFrontendTypes'
-import utils from '../../utils/utils'
+import { TextareaArgs } from '../../utils/govukFrontendTypes'
+import RiskView from '../shared/riskView'
 
 export default class RiskInformationView {
   constructor(private readonly presenter: RiskInformationPresenter) {}
+
+  private riskView = new RiskView(this.presenter.riskPresenter)
 
   private readonly errorSummaryArgs = ViewUtils.govukErrorSummaryArgs(this.presenter.errorSummary)
 
@@ -13,69 +15,14 @@ export default class RiskInformationView {
       name: 'additional-risk-information',
       id: 'additional-risk-information',
       label: {
-        text: this.presenter.text.additionalRiskInformation.label,
-        classes: this.presenter.text.additionalRiskInformation.labelClasses,
+        text: this.presenter.additionalRiskInformation.label,
+        classes: this.presenter.additionalRiskInformation.labelClasses,
       },
       value: this.presenter.fields.additionalRiskInformation,
       hint: {
-        text: this.presenter.text.additionalRiskInformation.hint,
+        text: this.presenter.additionalRiskInformation.hint,
       },
-      errorMessage: ViewUtils.govukErrorMessage(this.presenter.text.additionalRiskInformation.errorMessage),
-    }
-  }
-
-  roshAnalysisTableArgs(tagMacro: (args: TagArgs) => string): TableArgs {
-    return {
-      head: this.presenter.roshAnalysisHeaders.map((header: string) => {
-        return { text: header }
-      }),
-      rows: this.presenter.roshAnalysisRows.map(row => {
-        return [
-          { text: utils.convertToProperCase(row.riskTo) },
-          {
-            text: tagMacro({
-              text: row.riskScore.replace('_', ' '),
-              classes: this.tagClassForRiskScore(row.riskScore),
-            }),
-          },
-        ]
-      }),
-    }
-  }
-
-  private tagClassForRiskScore(riskScore: string): string {
-    switch (riskScore) {
-      case 'LOW':
-        return 'govuk-tag--green'
-      case 'MEDIUM':
-        return 'govuk-tag--blue'
-      case 'HIGH':
-        return 'govuk-tag--red'
-      case 'VERY_HIGH':
-        return 'govuk-tag--purple'
-      default:
-        return ''
-    }
-  }
-
-  get riskLevelDetailsArgs(): DetailsArgs {
-    return {
-      summaryText: 'Definitions of risk levels',
-      html: `
-            <ul class="govuk-list govuk-list--bullet">
-              <li>
-                <strong>Low</strong> - Current evidence does not indicate likelihood of causing serious harm.
-              </li>
-              <li>
-                <strong>Medium</strong> - There are identifiable indicators of risk of serious harm. The person has the potential to cause serious harm but is unlikely to do so unless there is a change in circumstances.
-              </li>
-              <li>
-                <strong>High</strong> - There are identifiable indicators of risk of serious harm. The potential event could happen at any time and the impact would be serious.
-              </li>
-              <li>
-                <strong>Very high</strong> - There is an imminent risk of serious harm. The potential event is more likely than not to happen imminently and the impact would be serious.
-              </li>
-            </ul>`,
+      errorMessage: ViewUtils.govukErrorMessage(this.presenter.additionalRiskInformation.errorMessage),
     }
   }
 
@@ -86,8 +33,8 @@ export default class RiskInformationView {
         presenter: this.presenter,
         errorSummaryArgs: this.errorSummaryArgs,
         additionalRiskInformationTextareaArgs: this.additionalRiskInformationTextareaArgs,
-        roshAnalysisTableArgs: this.roshAnalysisTableArgs.bind(this),
-        riskLevelDetailsArgs: this.riskLevelDetailsArgs,
+        roshAnalysisTableArgs: this.riskView.roshAnalysisTableArgs.bind(this.riskView),
+        riskLevelDetailsArgs: this.riskView.riskLevelDetailsArgs,
       },
     ]
   }
