+++ server/views/referrals/riskInformation.njk
@@ -23,7 +23,7 @@
 
       <h1 class="govuk-heading-xl">{{ presenter.text.title }}</h1>
 
-      {% if presenter.riskSummaryEnabled %}
+      {% if presenter.riskPresenter.riskSummaryEnabled %}
 
         <p class="govuk-body">This shows the current ROSH levels and risk to self information from OASys. You can add information for the service provider.</p>
